+++ src/main/kotlin/pt/isel/ls/service/entities/Route.kt
@@ -16,7 +16,7 @@ data class Route(
     val id: RouteID,
     val startLocation: String,
     val endLocation: String,
-    val distance: Double,
+    val distance: Float,
     val user: UserID
 ) {
     companion object {
